Change project license from GPLv2 to MIT
This commit updates the licensing of the project from General Public License version 2 (GPLv2) to the MIT license. This change is made to provide more flexibility in how the code can be used and redistributed. All references to the previous license in the source files and documentation have been updated to reflect this change.